Fix handlebars language loader
Handlebars shares a worker with html in the same way that
javascript/typescript and css/less/scss share workers. However, the
handlebars/html association was missing from the createLoaderRules
function. This was causing a bad network request when the language was
set to handlebars in the monaco-editor due to trying to load an
undefined web worker.

This change associates the handlebars and html workerPaths to prevent
these issues.