Add asserts, bigint, and unknown to TypeScript keywords
Also, while I'm in here, consolidate typeKeywords into keywords in order to make it easier to keep this in sync with TypeScript going forward. I'm assuming that typeKeywords is not part of Monaco's public API that third-party integrations are supposed to depend on; if it is, then I can undo this.

Fixes: microsoft/monaco-editor#1978