[webpack readme] Add how to get languages/features
I had to do a lot of digging in the source code just to know what strings I can use in the features list. Adding this allows users to find out for themselves. I considered adding the list here but I don't know if it might change depending on your build.